dividing notify function from createPost file and adding it to updatePost too
